requirements.txt
Django>=2.0,<3.0
gunicorn==19.9.0
django-mysql>=2.2.0
djangorestframework==3.11.0
requests==2.23.0